[#21][refactor]User를 user 패키지로 이동
