+++ src/main/java/dev/idion/bladitodo/domain/user/User.java
@@ -1,5 +1,6 @@
-package dev.idion.bladitodo.domain.entity;
+package dev.idion.bladitodo.domain.user;
 
+import dev.idion.bladitodo.domain.entity.BaseTimeEntity;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
